refactoring file locations
